csrf wrong

+++ tudo/app/_serverActions/logout.tsx
@@ -5,12 +5,19 @@ import { cookies } from "next/headers";
 const LogOut = async() => {
     try{
         let cookieStore = await cookies();
-
-        cookieStore.delete('access_token');
-        cookieStore.delete('refresh_token');
-        redirect("/singIn")
+        const access = cookieStore.get('access_token')
+        const refresh = cookieStore.get('refresh_token')
+        
+        if (!refresh || !access){
+            redirect("/singIn")
+        }else{
+            cookieStore.delete('access_token')
+            cookieStore.delete('refresh_token')
+            redirect("/singIn")
+        }
+        
     }finally{
-
+        redirect("/singIn")
     }
 }
 export default LogOut;
